Fix L0_resnet_dali_sanity for Dali 0.2
The new changes in daliop's API require that 'height', 'width' and 'batch_size' are supplied as
members of a tuple as an argument 'shape'.

Further changes inclue removing TF casts, as they interfere with the (new) polymorphic API of
daliop, which allows for explicitly selecting the type of 'samples_type' and 'labels_type'.

Signed-off-by: Pablo Ribalta <pribalta@nvidia.com>