Run self tests without rebuild.
Cut down L1_self_test to the python/kernel_test. Test commands for
these and xla tests are captured during the build of the base
container. This increases the time for the initial build by a few
minutes, but greatly decreases the amount of time spent running the
tests.

The old build-and-test scripts have been relocated to an L3 test.